# commodities_api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AtlasMarkets — Pollux",
    "version": "1.0.0",
    "description": "Commodities market intelligence for AI agents. x402-protected on Base mainnet.",
    "x-guidance": "AtlasMarkets Pollux provides commodities market intelligence for AI agents. Routes: GET /api/pollux/signals ($0.05), GET /api/pollux/forecast?symbol=Gold ($0.05), GET /api/pollux/risk ($0.02), GET /api/pollux/preflight?symbol=Gold ($0.05), GET /api/pollux/history?symbol=Gold ($0.05), GET /api/pollux/audit?decision_id=X ($0.07), POST /api/pollux/decision ($0.15). All routes return real-time commodities data.",
    "contact": {
      "email": "max.sadikovic@gmail.com"
    }
  },
  "servers": [
    {
      "url": "https://atlasmarkets-pollux.onrender.com",
      "description": "Production"
    }
  ],
  "paths": {
    "/api/pollux/signals": {
      "get": {
        "operationId": "getPolluxSignals",
        "tags": ["Commodities", "Intelligence", "Signals"],
        "summary": "Current commodities market signals",
        "x-payment-info": {
          "price": {"mode": "fixed", "currency": "USD", "amount": "0.050000"},
          "protocols": [{"x402": {}}]
        },
        "parameters": [
          {
            "name": "timeframe",
            "in": "query",
            "required": false,
            "schema": {"type": "string", "default": "1d"},
            "description": "Time window e.g. 1h, 1d, 1w"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignalsResponse"
                }
              }
            }
          },
          "402": {"description": "Payment Required"}
        }
      }
    },
    "/api/pollux/decision": {
      "post": {
        "operationId": "postPolluxDecision",
        "tags": ["Commodities", "Decision"],
        "summary": "Make a commodities market decision",
        "x-payment-info": {
          "price": {"mode": "fixed", "currency": "USD", "amount": "0.150000"},
          "protocols": [{"x402": {}}]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Decision outcome",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionResponse"
                }
              }
            }
          },
          "402": {"description": "Payment Required"}
        }
      }
    },
    "/api/pollux/audit": {
      "get": {
        "operationId": "getPolluxAudit",
        "tags": ["Commodities", "Audit"],
        "summary": "Audit a prior decision",
        "x-payment-info": {
          "price": {"mode": "fixed", "currency": "USD", "amount": "0.070000"},
          "protocols": [{"x402": {}}]
        },
        "parameters": [
          {
            "name": "decision_id",
            "in": "query",
            "required": true,
            "schema": {"type": "string"},
            "description": "Decision ID to audit"
          },
          {
            "name": "window",
            "in": "query",
            "required": false,
            "schema": {"type": "string", "default": "1h"},
            "description": "Evaluation window e.g. 1h"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditResponse"
                }
              }
            }
          },
          "402": {"description": "Payment Required"}
        }
      }
    },
    "/api/pollux/forecast": {
      "get": {
        "operationId": "getPolluxForecast",
        "tags": ["Commodities", "Forecast"],
        "summary": "Conformally-calibrated price forecast",
        "x-payment-info": {
          "price": {"mode": "fixed", "currency": "USD", "amount": "0.050000"},
          "protocols": [{"x402": {}}]
        },
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": false,
            "schema": {"type": "string", "default": "Gold (XAU/USD)"},
            "description": "Commodity name e.g. Gold (XAU/USD)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForecastResponse"
                }
              }
            }
          },
          "402": {"description": "Payment Required"}
        }
      }
    },
    "/api/pollux/risk": {
      "get": {
        "operationId": "getPolluxRisk",
        "tags": ["Commodities", "Risk"],
        "summary": "Current market risk state",
        "x-payment-info": {
          "price": {"mode": "fixed", "currency": "USD", "amount": "0.020000"},
          "protocols": [{"x402": {}}]
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskResponse"
                }
              }
            }
          },
          "402": {"description": "Payment Required"}
        }
      }
    },
    "/api/pollux/preflight": {
      "get": {
        "operationId": "getPolluxPreflight",
        "tags": ["Commodities", "Preflight"],
        "summary": "Pre-decision conditions check",
        "x-payment-info": {
          "price": {"mode": "fixed", "currency": "USD", "amount": "0.050000"},
          "protocols": [{"x402": {}}]
        },
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": false,
            "schema": {"type": "string", "default": "Gold (XAU/USD)"},
            "description": "Commodity name e.g. Gold (XAU/USD)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "symbol": {"type": "string"},
                    "ts": {"type": "string"},
                    "can_decide": {"type": "boolean"},
                    "cooldown_active": {"type": "boolean"},
                    "market_state": {"type": "string"},
                    "price": {"type": "number"},
                    "volatility": {"type": "string"},
                    "warnings": {"type": "array", "items": {"type": "string"}},
                    "data_freshness": {"type": "string"}
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "402": {"description": "Payment Required"}
        }
      }
    },
    "/api/pollux/history": {
      "get": {
        "operationId": "getPolluxHistory",
        "tags": ["Commodities", "History"],
        "summary": "Recent context history",
        "x-payment-info": {
          "price": {"mode": "fixed", "currency": "USD", "amount": "0.050000"},
          "protocols": [{"x402": {}}]
        },
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": false,
            "schema": {"type": "string", "default": "Gold (XAU/USD)"},
            "description": "Commodity name e.g. Gold (XAU/USD)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {"type": "integer", "default": 10},
            "description": "Max entries"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "symbol": {"type": "string"},
                    "count": {"type": "integer"},
                    "history": {"type": "array"},
                    "data_freshness": {"type": "string"}
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "402": {"description": "Payment Required"}
        }
      }
    },
    "/health": {
      "get": {
        "operationId": "getHealth",
        "tags": ["Utility"],
        "summary": "Health check",
        "security": [],
        "responses": {
          "200": {
            "description": "Service is healthy",
            "content": {
              "application/json": {
                "schema": {"type": "object"}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "siwx": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "SIWX",
        "description": "Sign-In with X (SIWX) wallet authentication"
      }
    },
    "schemas": {
      "SignalsResponse": {
        "type": "object",
        "properties": {
          "ts": {"type": "string"},
          "timeframe": {"type": "string"},
          "regime": {"type": "string"},
          "signals": {"type": "object"},
          "top_k": {"type": "array", "items": {"type": "string"}},
          "signal_age_hours": {"type": "number"},
          "data_freshness": {"type": "string"}
        },
        "additionalProperties": false
      },
      "DecisionRequest": {
        "type": "object",
        "required": ["symbol"],
        "properties": {
          "symbol": {"type": "string", "description": "Commodity name e.g. Gold (XAU/USD)"}
        },
        "additionalProperties": false
      },
      "DecisionResponse": {
        "type": "object",
        "properties": {
          "decision_id": {"type": "string"},
          "symbol": {"type": "string"},
          "suggested_action": {"type": "string"},
          "confidence": {"type": "number"},
          "certainty": {"type": "string"},
          "directional_edge": {"type": "string"},
          "raw_signal": {"type": "number"},
          "regime": {"type": "string"},
          "risk_level": {"type": "string"},
          "data_freshness": {"type": "string"},
          "next_step": {"type": "object"}
        },
        "additionalProperties": false
      },
      "AuditResponse": {
        "type": "object",
        "properties": {
          "decision_id": {"type": "string"},
          "symbol": {"type": "string"},
          "suggested_action": {"type": "string"},
          "confidence": {"type": "number"},
          "evaluation_window": {"type": "string"},
          "prices": {"type": "object"},
          "outcome": {"type": "object"}
        },
        "additionalProperties": false
      },
      "ForecastResponse": {
        "type": "object",
        "properties": {
          "symbol": {"type": "string"},
          "ts": {"type": "string"},
          "regime": {"type": "string"},
          "forecast": {"type": "object"},
          "data_freshness": {"type": "string"}
        },
        "additionalProperties": false
      },
      "RiskResponse": {
        "type": "object",
        "properties": {
          "ts": {"type": "string"},
          "regime": {"type": "string"},
          "risk_level": {"type": "string"},
          "risk_factors": {"type": "array", "items": {"type": "string"}},
          "cooldown_active": {"type": "boolean"},
          "data_freshness": {"type": "string"}
        },
        "additionalProperties": false
      }
    }
  },
  "x-x402": {
    "network": "eip155:8453",
    "payTo": "0x8eB96caA976De43027FEf619c4D24F6679486277",
    "facilitator": "https://x402.xyz/facilitate",
    "extensions": {
      "bazaar": {
        "status": "live",
        "purpose": "AtlasMarkets Pollux commodities market intelligence for AI agents."
      }
    }
  }
}